# src/main/java/kr/hhplus/be/server/domain/concert/schedule/entity/ConcertSchedule.java
package kr.hhplus.be.server.domain.concert.schedule.entity;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.AccessLevel;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

import java.time.LocalDate;

@Getter
@NoArgsConstructor(access = AccessLevel.PROTECTED)
@Entity
public class ConcertSchedule {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private Long concertId;

    private LocalDate concertDate;

    @Builder
    private ConcertSchedule(Long concertId, LocalDate concertDate) {
        this.concertId = concertId;
        this.concertDate = concertDate;
    }

}